update animations delay
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@1963 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/system/settings.cpp
@@ -631,7 +631,7 @@ bool Settings::CDMusic(void) const { return opt_global.Modes(GLOBAL_MUSIC_CD | G
 /* return animation */
 u8   Settings::Animation(void) const { return animation; }
 
-u8   Settings::Performance(void) const { return performance; }
+u16  Settings::Performance(void) const { return performance; }
 
 /* return full screen */
 bool Settings::FullScreen(void) const { return opt_global.Modes(GLOBAL_FULLSCREEN); }
